// src/app/shared/enums/roles.ts
export enum Roles {
    super_admin = 'super_admin',
    admin = 'admin',
    api_partner = 'api_partner',
    super_distributor = 'super_distributor',
    distributor = 'distributor',
    merchant = 'merchant',
    retailer = 'retailer',
    customer = 'customer',
    employee = 'employee',
    guest = 'guest',
    sub_admin = 'sub_admin'
}